==> unswnb15/1.2.geracao_dataset_unsw2015_full_with_preprocessing.ipynb <==
import pandas as pd
import numpy as np

# Carregando dataset UNSW completo

df_unsw1 = pd.read_csv('/media/kelson/Maxtor/Trabalho/UFU/artigo_pre-processamento/dataset_unsw2015_csv/UNSW-NB15_1.csv', low_memory=False)
df_unsw1.columns = [nome_coluna.strip().replace(' ', '_').replace('/', '_') for nome_coluna in df_unsw1.columns]
df_unsw2 = pd.read_csv('/media/kelson/Maxtor/Trabalho/UFU/artigo_pre-processamento/dataset_unsw2015_csv/UNSW-NB15_2.csv', low_memory=False)
df_unsw3 = pd.read_csv('/media/kelson/Maxtor/Trabalho/UFU/artigo_pre-processamento/dataset_unsw2015_csv/UNSW-NB15_3.csv', low_memory=False)
df_unsw4 = pd.read_csv('/media/kelson/Maxtor/Trabalho/UFU/artigo_pre-processamento/dataset_unsw2015_csv/UNSW-NB15_4.csv', low_memory=False)

df_unsw1.columns

# os nomes das colunas dos datasets 2, 3 e 4 estão faltando
df_unsw3

# Análise e tratamento das colunas (renomeando)

# criando as colunas do dataset 2
df_unsw2 = df_unsw2.rename(columns={'59.166.0.0':'srcip',
                          '6055':'sport',
                          '149.171.126.5':'dstip',
                          '54145':'dsport',
                          'tcp':'proto',
                          'FIN':'state',
                          '0.072974':'dur',
                          '4238':'sbytes',
                          '60788':'dbytes',
                          '31':'sttl',
                          '29':'dttl',
                          '7':'sloss',
                          '30':'dloss',
                          '-':'service',
                          '458245.4375':'Sload',
                          '6571546.5':'Dload',
                          '72':'Spkts',
                          '72.1':'Dpkts',
                          '255':'swin',
                          '255.1':'dwin',
                          '1003293149':'stcpb',
                          '1003585034':'dtcpb',
                          '59':'smeansz',
                          '844':'dmeansz',
                          '0':'trans_depth',
                          '0.1':'res_bdy_len',
                          '62.04531':'Sjit',
                          '61.899776':'Djit',
                          '1421955842':'Stime',
                          '1421955842.1':'Ltime',
                          '1.02269':'Sintpkt',
                          '0.997042':'Dintpkt',
                          '0.002317':'tcprtt',
                          '0.002173':'synack',
                          '0.000144':'ackdat',
                          '0.2':'is_sm_ips_ports',
                          '0.3':'ct_state_ttl',
                          '0.4':'ct_flw_http_mthd',
                          '0.5':'is_ftp_login',
                          '0.6':'ct_ftp_cmd',
                          '13':'ct_srv_src',
                          '13.1':'ct_srv_dst',
                          '6':'ct_dst_ltm',
                          '7.1':'ct_src__ltm',
                          '1':'ct_src_dport_ltm',
                          '1.1':'ct_dst_sport_ltm',
                          '2':'ct_dst_src_ltm',
                          'Unnamed: 47':'attack_cat',
                          '0.7':'Label'})
df_unsw2

# criando as colunas do dataset 3
df_unsw3 = df_unsw3.rename(columns={'59.166.0.1':'srcip',
                          '18247':'sport',
                          '149.171.126.4':'dstip',
                          '7662':'dsport',
                          'tcp':'proto',
                          'FIN':'state',
                          '0.119596':'dur',
                          '4550':'sbytes',
                          '68342':'dbytes',
                          '31':'sttl',
                          '29':'dttl',
                          '7':'sloss',
                          '33':'dloss',
                          '-':'service',
                          '300478.2813':'Sload',
                          '4514398.5':'Dload',
                          '78':'Spkts',
                          '80':'Dpkts',
                          '255':'swin',
                          '255.1':'dwin',
                          '1818376620':'stcpb',
                          '1818657356':'dtcpb',
                          '58':'smeansz',
                          '854':'dmeansz',
                          '0':'trans_depth',
                          '0.1':'res_bdy_len',
                          '87.137592':'Sjit',
                          '85.643619':'Djit',
                          '1424231129':'Stime',
                          '1424231129.1':'Ltime',
                          '1.549156':'Sintpkt',
                          '1.50719':'Dintpkt',
                          '0.000644':'tcprtt',
                          '0.000521':'synack',
                          '0.000123':'ackdat',
                          '0.2':'is_sm_ips_ports',
                          '0.3':'ct_state_ttl',
                          'Unnamed: 37':'ct_flw_http_mthd',
                          'Unnamed: 38':'is_ftp_login',
                          ' ':'ct_ftp_cmd',
                          '6':'ct_srv_src',
                          '2':'ct_srv_dst',
                          '2.1':'ct_dst_ltm',
                          '5':'ct_src__ltm',
                          '1':'ct_src_dport_ltm',
                          '1.1':'ct_dst_sport_ltm',
                          '2.2':'ct_dst_src_ltm',
                          'Unnamed: 47':'attack_cat',
                          '0.4':'Label'})
df_unsw3

# criando as colunas do dataset 4
df_unsw4 = df_unsw4.rename(columns={'59.166.0.9':'srcip',
                          '7045':'sport',
                          '149.171.126.7':'dstip',
                          '25':'dsport',
                          'tcp':'proto',
                          'FIN':'state',
                          '0.201886':'dur',
                          '37552':'sbytes',
                          '3380':'dbytes',
                          '31':'sttl',
                          '29':'dttl',
                          '18':'sloss',
                          '8':'dloss',
                          'smtp':'service',
                          '1459437.5':'Sload',
                          '130766.8672':'Dload',
                          '52':'Spkts',
                          '42':'Dpkts',
                          '255':'swin',
                          '255.1':'dwin',
                          '1422136554':'stcpb',
                          '3572668484':'dtcpb',
                          '722':'smeansz',
                          '80':'dmeansz',
                          '0':'trans_depth',
                          '0.1':'res_bdy_len',
                          '456.043567':'Sjit',
                          '15.530109':'Djit',
                          '1424250009':'Stime',
                          '1424250009.1':'Ltime',
                          '3.943843':'Sintpkt',
                          '4.912488':'Dintpkt',
                          '0.00059':'tcprtt',
                          '0.000473':'synack',
                          '0.000117':'ackdat',
                          '0.2':'is_sm_ips_ports',
                          '0.3':'ct_state_ttl',
                          'Unnamed: 37':'ct_flw_http_mthd',
                          'Unnamed: 38':'is_ftp_login',
                          ' ':'ct_ftp_cmd',
                          '2':'ct_srv_src',
                          '2.1':'ct_srv_dst',
                          '7':'ct_dst_ltm',
                          '4':'ct_src__ltm',
                          '1':'ct_src_dport_ltm',
                          '1.1':'ct_dst_sport_ltm',
                          '3':'ct_dst_src_ltm',
                          'Unnamed: 47':'attack_cat',
                          '0.4':'Label'})
df_unsw4

# juntandos os datasets 1, 2, 3 e 4
df_unsw = pd.concat([df_unsw1, df_unsw2, df_unsw3, df_unsw4])

# os índices estão duplicados
df_unsw.loc[[100]]

# reseta os índices e faz a verificação
df_unsw.reset_index(drop=True, inplace=True)
df_unsw.loc[[100]]

#pd.set_option("display.max_columns", None)
#print(df_unsw.head(100))
#pd.set_option("display.max_columns", False)

df_unsw.ct_flw_http_mthd.value_counts()

# visualizando os valores nulos
df_unsw.isnull().sum()

# consulta quantidade de registro de uma categoria dentro da coluna específica
#df_unsw.ct_flw_http_mthd.value_counts()

# retorna todas as linhas com NaN da coluna específica
# ct_flw_http_mthd
# is_ftp_login
# attack_cat
df_unsw[df_unsw['ct_flw_http_mthd'].isnull()].head(100)

# substituir valores NaN da coluna específica
df_unsw['attack_cat'].fillna('Benign', inplace=True)
df_unsw['is_ftp_login'].fillna(3, inplace=True)
df_unsw['ct_flw_http_mthd'].fillna(7, inplace=True)

#df_unsw.isnull().sum()
# ct_flw_http_mthd
# is_ftp_login
# attack_cat
df_unsw.is_ftp_login.value_counts()

# uniificando os ataques duplicados
df_unsw.loc[df_unsw.attack_cat == 'Backdoors', df_unsw.attack_cat.name] = 'Backdoor'
df_unsw.loc[df_unsw.attack_cat == ' Reconnaissance ', df_unsw.attack_cat.name] = 'Reconnaissance'
df_unsw.loc[df_unsw.attack_cat == ' Shellcode ', df_unsw.attack_cat.name] = 'Shellcode'
df_unsw.loc[df_unsw.attack_cat == ' Fuzzers ', df_unsw.attack_cat.name] = 'Fuzzers'
df_unsw.loc[df_unsw.attack_cat == ' Fuzzers', df_unsw.attack_cat.name] = 'Fuzzers'

df_unsw.attack_cat.value_counts()

# FAZER APENAS NO PRÉ-PROCESSAMENTO
# colunas excluídas que não possuem influência na detecção
#df_unsw =  df_unsw.drop(['srcip', 'sport', 'dstip', 'dsport', 'Stime', 'Ltime'], axis='columns')
df_unsw =  df_unsw.drop(['srcip', 'dstip'], axis='columns')
df_unsw.shape

# alterando nome (-) da classe service para 'others'
df_unsw.loc[df_unsw.service == '-', df_unsw.service.name] = 'others'
# alterando nome ' ' da classe ct_ftp_cmd para '7'
df_unsw.loc[df_unsw.ct_ftp_cmd == ' ', df_unsw.ct_ftp_cmd.name] = '7'

df_unsw.service.value_counts()

pd.set_option('display.max_rows', None)
print(df_unsw.dtypes)
pd.set_option('display.max_rows', False)

# convertendo variáveis específicas de nominal para inteiro
df_unsw['ct_ftp_cmd'] = df_unsw['ct_ftp_cmd'].astype(int)
#df_unsw['srcip'] = df_unsw['srcip'].astype(float)

# verifica detalhes das variáveis
# variáveis nominais: #proto #state #service #ct_ftp_cmd
#df_unsw.shape
#df_unsw.describe()
#df_unsw.dtypes
#df_unsw.ct_ftp_cmd.value_counts()
#np.unique(df_unsw['sport'])
#len(np.unique(df_unsw.ct_ftp_cmd))
pd.set_option('display.max_rows', None)
print(df_unsw.dtypes)
pd.set_option('display.max_rows', False)

# exportar dataset unsw completo
df_unsw.to_csv('/media/kelson/Maxtor/Trabalho/doutorado/artigo_preprocessamento/artigo/dataset_unsw/df_unsw_full.csv', index=False)
