==> unswnb15/2.1.geracao_dataset_unsw2015_labelencoder_train_test_no_preprocessing.ipynb <==
import pandas as pd
import numpy as np
from sklearn.utils import shuffle

# Carregando dataset UNSW completo pre-processamento MINIMO

df_unsw_full = pd.read_csv('/media/kelson/Maxtor/Trabalho/doutorado/artigo_preprocessamento/artigo/dataset_unsw/df_unsw_full.csv', low_memory=False)

df_unsw_full

# visualizando os valores nulos
df_unsw_full.isnull().sum()

# verifica detalhes das variáveis
# variáveis nominais: #proto #state #service #ct_ftp_cmd
#df_unsw_full.shape
#df_unsw_full.describe()
#df_unsw_full.dtypes
#df_unsw_full.ct_ftp_cmd.value_counts()
#np.unique(df_unsw_full['sport'])
#len(np.unique(df_unsw_full.ct_ftp_cmd))
pd.set_option('display.max_rows', None)
print(df_unsw_full.attack_cat.value_counts())
pd.set_option('display.max_rows', False)

# Pré-processamento com Condificação (LabelEncoder)

from sklearn.preprocessing import LabelEncoder

pd.set_option('display.max_rows', None)
print(df_unsw_full.dtypes)
pd.set_option('display.max_rows', False)

df_unsw_full.head()

# cria variáveis recebendo a função LabelEncoder
label_encoder_srcip = LabelEncoder()
label_encoder_sport = LabelEncoder()
label_encoder_dstip = LabelEncoder()
label_encoder_dsport = LabelEncoder()
label_encoder_proto  = LabelEncoder()
label_encoder_state = LabelEncoder()
label_encoder_service = LabelEncoder()

# faz a codificação das variáveis nominais (colunas) específicas
#srcip #sport #dstip #dsport #proto #state #service
df_unsw_full['srcip'] = label_encoder_srcip.fit_transform(df_unsw_full['srcip'])
df_unsw_full['sport'] = label_encoder_sport.fit_transform(df_unsw_full['sport'])
df_unsw_full['dstip'] = label_encoder_dstip.fit_transform(df_unsw_full['dstip'])
df_unsw_full['dsport'] = label_encoder_dsport.fit_transform(df_unsw_full['dsport'])
df_unsw_full['proto'] = label_encoder_proto.fit_transform(df_unsw_full['proto'])
df_unsw_full['state'] = label_encoder_state.fit_transform(df_unsw_full['state'])
df_unsw_full['service'] = label_encoder_service.fit_transform(df_unsw_full['service'])

df_unsw_full.head()

pd.set_option('display.max_rows', None)
print(df_unsw_full.dtypes)
pd.set_option('display.max_rows', False)

# Divisão dataset full em Train (50%) e Test - Base-Line (50%)

# separação dos tipos de ataque (coluna attack_cat)
df_benign = df_unsw_full[df_unsw_full.attack_cat == 'Benign']
df_generic = df_unsw_full[df_unsw_full.attack_cat == 'Generic']
df_exploits = df_unsw_full[df_unsw_full.attack_cat == 'Exploits']
df_fuzzers = df_unsw_full[df_unsw_full.attack_cat == 'Fuzzers']
df_dos = df_unsw_full[df_unsw_full.attack_cat == 'DoS']
df_reconnaissance = df_unsw_full[df_unsw_full.attack_cat == 'Reconnaissance']
df_analysis = df_unsw_full[df_unsw_full.attack_cat == 'Analysis']
df_backdoor = df_unsw_full[df_unsw_full.attack_cat == 'Backdoor']
df_shellcode = df_unsw_full[df_unsw_full.attack_cat == 'Shellcode']
df_worms = df_unsw_full[df_unsw_full.attack_cat == 'Worms']

# dimensão do dataframe fragmentado
print(df_benign.shape)
print(df_generic.shape)
print(df_exploits.shape)
print(df_fuzzers.shape)
print(df_dos.shape)
print(df_reconnaissance.shape)
print(df_analysis.shape)
print(df_backdoor.shape)
print(df_shellcode.shape)
print(df_worms.shape)

# reseta os índices
df_benign.reset_index(drop=True, inplace=True)
df_generic.reset_index(drop=True, inplace=True)
df_exploits.reset_index(drop=True, inplace=True)
df_fuzzers.reset_index(drop=True, inplace=True)
df_dos.reset_index(drop=True, inplace=True)
df_reconnaissance.reset_index(drop=True, inplace=True)
df_analysis.reset_index(drop=True, inplace=True)
df_backdoor.reset_index(drop=True, inplace=True)
df_shellcode.reset_index(drop=True, inplace=True)
df_worms.reset_index(drop=True, inplace=True)

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_benign50 = df_benign.iloc[:1109380,:]
df2_benign50 = df_benign.iloc[1109381:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_generic50 = df_generic.iloc[:107740,:]
df2_generic50 = df_generic.iloc[107741:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_exploits50 = df_exploits.iloc[:22262,:]
df2_exploits50 = df_exploits.iloc[22263:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_fuzzers50 = df_fuzzers.iloc[:12122,:]
df2_fuzzers50 = df_fuzzers.iloc[12123:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_dos50 = df_dos.iloc[:8176,:]
df2_dos50 = df_dos.iloc[8177:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_reconnaissance50 = df_reconnaissance.iloc[:6993,:]
df2_reconnaissance50 = df_reconnaissance.iloc[6994:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_analysis50 = df_analysis.iloc[:1338,:]
df2_analysis50 = df_analysis.iloc[1339:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_backdoor50 = df_backdoor.iloc[:1164,:]
df2_backdoor50 = df_backdoor.iloc[1165:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_shellcode50 = df_shellcode.iloc[:755,:]
df2_shellcode50 = df_shellcode.iloc[756:,:]

# divisão dos dataframe fragmentados por linha (50 a 50)
df1_worms50 = df_worms.iloc[:86,:]
df2_worms50 = df_worms.iloc[87:,:]

# juntando os dataframes train50
df_unsw_train50 = pd.concat([df1_benign50, df1_generic50, df1_exploits50, df1_fuzzers50, df1_dos50, 
                             df1_reconnaissance50, df1_analysis50, df1_backdoor50, df1_shellcode50, df1_worms50])
# embaralha do dataframe 
df_unsw_train50 = shuffle(df_unsw_train50)
# reseta os indices
df_unsw_train50.reset_index(drop=True, inplace=True)
# verifica a dimensão
df_unsw_train50.shape

# juntando os dataframes train50
df_unsw_test50 = pd.concat([df2_benign50, df2_generic50, df2_exploits50, df2_fuzzers50, df2_dos50, 
                             df2_reconnaissance50, df2_analysis50, df2_backdoor50, df2_shellcode50, df2_worms50])
# embaralha do dataframe 
df_unsw_test50 = shuffle(df_unsw_test50)
# reseta os indices
df_unsw_test50.reset_index(drop=True, inplace=True)
# verifica a dimensão
df_unsw_test50.shape

df_unsw_test50.tail(500)

# exportando dataset unsw train50
df_unsw_train50.to_csv('/media/kelson/Maxtor/Trabalho/doutorado/artigo_preprocessamento/artigo/dataset_unsw/df_unsw_train50.csv', index=False)

# exportando dataset unsw test50
df_unsw_test50.to_csv('/media/kelson/Maxtor/Trabalho/doutorado/artigo_preprocessamento/artigo/dataset_unsw/df_unsw_test50.csv', index=False)
