==> unswnb15/4.2.division_features_train50_test50_unsw2015_with_preprocessing_standadscaler.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import plotly.express as px
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report
from yellowbrick.classifier import ConfusionMatrix

# Carregando dataset UNSW train pre-processamento full

df_unsw_train50 = pd.read_csv('/media/kelson/DataCenter/UFU/artigo/unsw2105/df_unsw_train50_with_preprocessing_getdummies.csv', low_memory=False)

df_unsw_train50.head()

# Divisão entre variáveis previsoras (x) e classe (y)

# renomeando as variáveis classificadoras
#df_unsw_train50['Label_multiclass'] = df_unsw_train50['attack_cat']
#df_unsw_train50['Label_binary'] = df_unsw_train50['Label']

# apagando as variáveis classificadoras renomeadas
#df_unsw_train50.drop('attack_cat', axis = 1, inplace=True)
#df_unsw_train50.drop('Label', axis = 1, inplace=True)

#df_unsw_train50.head()

# atribuição dos valores para X e y
X = df_unsw_train50.iloc[:,:209]
y_multiclass = df_unsw_train50.Label_multiclass
y_binary = df_unsw_train50.Label_binary
print('Dimensões para X: ', X.shape)
print('Dimensões para y_multiclass: ', y_multiclass.shape)
print('Dimensões para y_binary: ', y_binary.shape)

X

y_multiclass

y_binary

y_binary.value_counts()

y_multiclass.value_counts()

# Divisão das variáveis entre treino e teste

from sklearn.model_selection import train_test_split

X_train50_train, X_train50_test, y_train50_multiclass_train, y_train50_multiclass_test, y_train50_binary_train, y_train50_binary_test = train_test_split(X, y_multiclass, y_binary, test_size = 0.3, random_state=42)

print('X treino = ', X_train50_train.shape)
print('X teste = ', X_train50_test.shape)
print('y multiclasse treino = ', y_train50_multiclass_train.shape)
print('y multiclasse teste = ', y_train50_multiclass_test.shape)
print('y binário treino = ', y_train50_binary_train.shape)
print('y binário teste = ', y_train50_binary_test.shape)

# Pré-processamento com Padronização (StandadScaler)

from sklearn.preprocessing import StandardScaler
# a padronização e a normalização devem ser aplicadas somente nos dados de treino (X_train)
# esse método evita o vazamento de informação, evitando que o algoritmo já conheça os valores médios normalizados

X_train50_train.describe()

X_train50_train.head()

# mostrando valores máximos das variáveis #proto #state #dur #service
X_train50_train['dur'].max(), X_train50_train['sbytes'].max(), X_train50_train['dbytes'].max(), X_train50_train['sttl'].max(), X_train50_train['dttl'].max()

# mostrando valores mínimos das variáveis #proto #state #dur #service
X_train50_train['dur'].min(), X_train50_train['sbytes'].min(), X_train50_train['dbytes'].min(), X_train50_train['sttl'].min(), X_train50_train['dttl'].min()

type(X_train50_train)

# realizando a padronização
# deve ser aplicado o transform nos dados de test
scaler_unsw = StandardScaler()
X_train50_train = scaler_unsw.fit_transform(X_train50_train)

# O X_train50_train foi convertido para o formato numpy array durante a aplicação do scaler
type(X_train50_train)

# mostrando valores máximos das variáveis #proto #state #dur #service
X_train50_train[:,0].max(), X_train50_train[:,1].max(), X_train50_train[:,2].max(), X_train50_train[:,3].max(), X_train50_train[:,4].max()

# mostrando valores máximos das variáveis #proto #state #dur #service
X_train50_train[:,0].min(), X_train50_train[:,1].min(), X_train50_train[:,2].min(), X_train50_train[:,3].min(),  X_train50_train[:,4].min()

# atribuindo novo dataframe somente para visualizar os novos valores do X_train50_train
df_X_train50_train = pd.DataFrame(scaler_unsw.transform(X_train50_train))
#df_X_train = pd.DataFrame(scaler_unsw.transform(X_train))

df_X_train50_train.describe()

### Aplicando o transform nos dados de teste (X_train50_test)

X_train50_test.describe()

# O X_train50_train foi convertido para o formato numpy array durante a aplicação do scaler
type(X_train50_test)

# mostrando valores máximos das variáveis #proto #state #dur #service
X_train50_test['dur'].max(), X_train50_test['sbytes'].max(), X_train50_test['dbytes'].max(), X_train50_test['sttl'].max(), X_train50_test['dttl'].max()

# mostrando valores máximos das variáveis #proto #state #dur #service
X_train50_test['dur'].min(), X_train50_test['sbytes'].min(), X_train50_test['dbytes'].min(), X_train50_test['sttl'].min(), X_train50_test['dttl'].min()

# realizando a padronização
# deve ser aplicado somente o transform nos dados de test
X_train50_test = scaler_unsw.transform(X_train50_test)

# O X_train50_train foi convertido para o formato numpy array durante a aplicação do scaler
type(X_train50_test)

# mostrando valores máximos das variáveis #proto #state #dur #service
X_train50_test[:,0].max(), X_train50_test[:,1].max(), X_train50_test[:,2].max(), X_train50_test[:,3].max(), X_train50_test[:,4].max()

# mostrando valores mínimos das variáveis #proto #state #dur #service
X_train50_test[:,0].min(), X_train50_test[:,1].min(), X_train50_test[:,2].min(), X_train50_test[:,3].min(), X_train50_test[:,4].min()

# Salvando as variáveis X_train50_train, X_train50_test, y_train50_multiclass_train, y_train50_multiclass_test, y_train50_binary_train, y_train50_binary_test em arquivo único

import pickle

with open('df_unsw_train50_with_preprocessing_getdummies_standadscaler', mode = 'wb') as f:
    pickle.dump([X_train50_train, X_train50_test, y_train50_multiclass_train, y_train50_multiclass_test, y_train50_binary_train, y_train50_binary_test], f)

# Carregando dataset UNSW Test50 pre-processamento full

df_unsw_test50 = pd.read_csv('/media/kelson/DataCenter/UFU/artigo/unsw2105/df_unsw_test50_with_preprocessing_getdummies.csv', low_memory=False)

df_unsw_test50.head()

# Divisão entre variáveis previsoras (x) e classe (y) do Test50

# renomeando as variáveis classificadoras
#df_unsw_test50['Label_multiclass'] = df_unsw_test50['attack_cat']
#df_unsw_test50['Label_binary'] = df_unsw_test50['Label']

# apagando as variáveis classificadoras renomeadas
#df_unsw_test50.drop('attack_cat', axis = 1, inplace=True)
#df_unsw_test50.drop('Label', axis = 1, inplace=True)

df_unsw_test50.head()

# atribuição dos valores para X e y
X_test50 = df_unsw_test50.iloc[:,:209]
y_test50_multiclass = df_unsw_test50.Label_multiclass
y_test50_binary = df_unsw_test50.Label_binary
print('Dimensões para X: ', X_test50.shape)
print('Dimensões para y_multiclass: ', y_test50_multiclass.shape)
print('Dimensões para y_binary: ', y_test50_binary.shape)

X_test50

y_test50_multiclass

y_test50_binary

# Pré-processamento com Padronização (StandadScaler) - somente transform

from sklearn.preprocessing import StandardScaler
# a padronização e a normalização devem ser aplicadas somente nos dados de treino (X_train)
# esse método evita o vazamento de informação, evitando que o algoritmo já conheça os valores médios normalizados
# deve ser aplicado somente o transform nos dados de test

X_test50.describe()

# O X_test50 foi convertido para o formato numpy array durante a aplicação do scaler
type(X_test50)

# mostrando valores máximos das variáveis #proto #state #dur #service
X_test50['dur'].max(), X_test50['sbytes'].max(), X_test50['dbytes'].max(), X_test50['sttl'].max(), X_test50['dttl'].max()

# mostrando valores máximos das variáveis #proto #state #dur #service
X_test50['dur'].min(), X_test50['sbytes'].min(), X_test50['dbytes'].min(), X_test50['sttl'].min(), X_test50['dttl'].min()

# realizando a padronização
# deve ser aplicado somente o transform nos dados de test
X_test50 = scaler_unsw.transform(X_test50)

# O X_test foi convertido para o formato numpy array durante a aplicação do scaler
type(X_test50)

# mostrando valores máximos das variáveis #proto #state #dur #service
X_test50[:,0].max(), X_test50[:,1].max(), X_test50[:,2].max(), X_test50[:,3].max(), X_test50[:,4].max()

# mostrando valores mínimos das variáveis #proto #state #dur #service
X_test50[:,0].min(), X_test50[:,1].min(), X_test50[:,2].min(), X_test50[:,3].min(), X_test50[:,4].min()

# Salvando as variáveis X_test50, y_test50_multiclass,  y_test50_binary em arquivo único

with open('df_unsw_test50_with_preprocessing_getdummies_standadscaler', mode = 'wb') as f:
    pickle.dump([X_test50, y_test50_multiclass, y_test50_binary], f)
